Update Demographic data Analyzer
In this project, I analyzed demographic data using Pandas. The dataset of demographic data that was extracted from the 1994 Census database. 
The project is meant to answer the following questions:
You must use Pandas to answer the following questions:

    How many people of each race are represented in this dataset? This should be a Pandas series with race names as the index labels. (race column)
    What is the average age of men?
    What is the percentage of people who have a Bachelor's degree?
    What percentage of people with advanced education (Bachelors, Masters, or Doctorate) make more than 50K?
    What percentage of people without advanced education make more than 50K?
    What is the minimum number of hours a person works per week?
    What percentage of the people who work the minimum number of hours per week have a salary of more than 50K?
    What country has the highest percentage of people that earn >50K and what is that percentage?
    Identify the most popular occupation for those who earn >50K in India
All decimals are rounded to the nearest tenth.

Link to the data: https://freecodecam-boilerplate-n3tye9o24xe.ws-eu114.gitpod.io/

import pandas as pd